Fix clicking at beginning/end of sine wave
This was due to having high amplitude at the very beginning of the sample and the
computer speaker was clicking as the diaphragm quickly snapped from 0 to max amplitude.
The solution was to add a soft start/stop capability that ramps up the amplitude over
the first 10-20 samples.

Fixes Issue #2