Extract common parts of SineOscillator into AbstractOscillator
We want to create some new kinds of oscillators. SineOscillator already
contains a decent amount of structure that we can reuse for other kinds
of oscillators, so this diff refactors things by creating an AbstractOscillator
class that contains the core functionality of each kind of oscillator
and a SawtoothOscillator that currently functions exactly like a SineOscillator
but will be modified in a later commit.